Rename testConstants.ts to constants.ts
