+++ public/pages/RollupDetails/containers/RollupDetails/RollupDetails.test.tsx
@@ -26,7 +26,7 @@ import { ModalProvider, ModalRoot } from "../../../../components/Modal";
 import { BREADCRUMBS, ROUTES } from "../../../../utils/constants";
 import RollupDetails from "./RollupDetails";
 import { ServicesConsumer, ServicesContext } from "../../../../services";
-import { testRollup, testRollup2 } from "../../../../../test/testConstants";
+import { testRollup, testRollup2 } from "../../../../../test/constants";
 
 function renderRollupDetailsWithRouter(initialEntries = ["/"]) {
   return {
